initial error fixes
